Finished a Pathfinding movement highlighter.
It is a bit inefficient, though I think this will be the most difficult
calculation in the game, so the lag will be minor.  I still have to
make a method that pulls a specific path and converts it into a list of
commands e.g. ["up", "up", "right"].  Also need to add in collisions
with other Actors.